Add create contest functionality to user
Added the ability for users to create individual contests. I created
a 'contest' route and a 'contest/new' nested route. This route loads a
new contest form component. I added a save action to this and passes the
data back up to make the call to the backend. I had an issue when I set
the 'DS.attr(date)' for some reason the data would not get pulled that
way. I set the attr('string') and that is working for now.
Next step: Implement the cancel functionality to the new contest form.