Fix UI routing issue with index on contests
Was having a problem where when creating a new contest I was also
seeing that list being made in the list of contests which was still
being rendered on the page. I was able to solve this by adding an
index file to the contests page where all of the users current contests
are being rendered and then providing a link to the new contest route.