Allow edit form to render
I finally got the edit form to render properly. I think I was loading
a model I didn't need to load in the contest.edit route. I am now
able to render the the form with the information pertaining to the
associated contest. I will now add the actions to allow for updating the
contest information