Add submission to contest
I added submissions to contests so that the submissions
associated with the contest render when in the contest/id
route. I had to serialize out the ids on the backend for this
to render properly. I seeded all the data so I need to now
allow for the creation of new submissions.