Fix model to allow for route render
I deleted the model on the 'app/contest/submission/route.js' route
This model wasn't being passsed anything and as a result it
was jamming up the route from loading. Removing the
model hook stopped this from happening. I am also now
able to render the route as a nested route.